third_party/WebKit/LayoutTests/platform/win/fast/spatial-navigation/snav-two-elements-one-line-expected.txt

AAAAAAAAAA AAAAAAAAAA AA AA AAAAAAAAAA AAAAAAAAAA


AAAAAAAAAA AAAAAAAAAA AA B AA AAAAAAAAAA AAAAAAAAAA
PASS successfullyParsed is true

TEST COMPLETE
FAIL gFocusedDocument.activeElement.getAttribute("id") should be e1. Was start.
FAIL gFocusedDocument.activeElement.getAttribute("id") should be e2. Was e1.
FAIL gFocusedDocument.activeElement.getAttribute("id") should be e3. Was e2.
FAIL gFocusedDocument.activeElement.getAttribute("id") should be e2. Was e1.
FAIL gFocusedDocument.activeElement.getAttribute("id") should be e1. Was start.
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
This test is testing that we can handle 2 inline elements in the same line.
